+++ xenon-common/src/main/java/com/vmware/xenon/services/common/ServiceHostManagementService.java
@@ -72,6 +72,9 @@ public class ServiceHostManagementService extends StatefulService {
     public static final String STAT_NAME_HTTP11_PENDING_OP_COUNT = "http11PendingOperationCount";
     public static final String STAT_NAME_HTTP2_PENDING_OP_COUNT = "http2PendingOperationCount";
 
+    public static final String STAT_NAME_EXECUTOR_QUEUE_DEPTH = "executorQueueDepth";
+    public static final String STAT_NAME_SCHEDULED_EXECUTOR_QUEUE_DEPTH = "scheduledExecutorQueueDepth";
+
     public static final String STAT_NAME_SERVICE_COUNT = "serviceCount";
     public static final String STAT_NAME_SERVICE_PAUSE_COUNT = "servicePauseCount";
     public static final String STAT_NAME_SERVICE_RESUME_COUNT = "serviceResumeCount";